Starts lasagne implementation of Andraw's code
Her code used TensorFlow to implement a CNN with which she was capable
of doing regression. My previous implementation using nolearn of a
similar network was not working (I assume, because of the loss
function it was minimizing: I tried changing her code to use the same
loss function as nolearn uses when `regression = True` in the
constructor of `nolearn.lasagne.NeuralNetwork` and the performance of
her code went down -- though I am not sure if this was the only
reason).

While I am trying to keep the nolearn implementation in
`cnn_nolearn.py`, this commit is the beginning of an effort to
implement her same network using Lasagne. The idea is to reproduce
her results with a different implementation.

 Changes to be committed:
	modified:   cnn_nolearn.py