A lot of restructuring...
Removes as much of MATLAB dependency as needed (or at least this is
my belief). Now the MNIST and CIFAR cross-validation folds are dumped
into files that are loaded by both the CDBN and the CAES. These will
also be loaded by the CNN that will be implemented in following
commits.

 Changes to be committed:
	modified:   CDBN
	modified:   auto_encoders/caes_nolearn.py
	renamed:    cnn_lasagne.py -> cnns/cnn_lasagne.py
	new file:   generate_cifar_folds.m
	new file:   generate_mnist_folds.m
	new file:   load_data_mnist_original.m
	new file:   run_autoencoders.sh
	new file:   run_cdbn_mnist.m
	modified:   run_networks.m