Implements CNN + minor fixes
Keras CNN is now working. It receives an input file with a dataset,
and an output folder, and fits a CNN with the data from the dataset.

Still need to output the results. I need the history of the loss
function, as well as the end accuracy in the test set. Then I will
finally be able to run this through all the crossvalidation folds.

 Changes to be committed:
	modified:   auto_encoders/caes_nolearn.py
	new file:   cnns/cnn_keras.py
	modified:   run_autoencoders.sh
	modified:   run_cdbn_mnist.m