Adapts CAES for CIFAR dataset
The Networks has the exact same architecture, only it receives a
bigger input and returns a bigger output.

There are also options for normalizing the input (CIFAR images values
are not between 0 and 1, but instead between 0 and 256. The
normalization simply divides the values by 256.0) and using 3 input
channels (CIFAR images are colored).

 Changes to be committed:
	modified:   auto_encoders/caes_nolearn.py