Prepare script for experiments (unfinished)
I still need to finish fixing `run_cnn_folds.sh` so that it uses all
variables that `run_cnns.sh` is passing to it. It should be able to
decide, for example, that a flag "--normalize" has to be passed to
`cnn_keras.py` when the dataset is CIFAR, or that "--use_relu" should
be passed to `cnn_keras.py` if `run_cnns.sh` passed `relu` to it.

 Changes to be committed:
	new file:   run_cnn_folds.sh
	modified:   run_cnns.sh